+++ build.gradle
@@ -8,7 +8,7 @@ repositories {
 }
 
 dependencies {
-    implementation 'com.github.woowacourse-projects:mission-utils:1.0.0'
+    implementation 'com.github.woowacourse-projects:mission-utils:1.2.0'
 }
 
 java {
