+++ src/main/java/bridge/InputView.java
@@ -1,6 +1,7 @@
 package bridge;
 
 import camp.nextstep.edu.missionutils.Console;
+import com.sun.security.jgss.GSSUtil;
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
@@ -19,6 +20,7 @@ public class InputView {
      * 사용자가 이동할 칸을 입력받는다.
      */
     public String readMoving() {
+        System.out.println();
         System.out.println("이동할 칸을 선택해주세요. (위: U, 아래: D)");
         return this.readUserInput();
     }
@@ -27,6 +29,7 @@ public class InputView {
      * 사용자가 게임을 다시 시도할지 종료할지 여부를 입력받는다.
      */
     public String readGameCommand() {
+        System.out.println();
         System.out.println("게임을 다시 시도할지 여부를 입력해주세요. (재시도: R, 종료: Q)");
         return this.readUserInput();
     }
