+++ src/main/java/bridge/domain/BridgeGame.java
@@ -3,6 +3,7 @@ package bridge.domain;
 import bridge.dtos.GameResultDto;
 import bridge.dtos.MapStateDto;
 import bridge.dtos.MapStateDto.SingleMapState;
+import bridge.enums.BridgeCell;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
@@ -18,11 +19,6 @@ public class BridgeGame {
         this.mapState = new MapState();
     }
 
-    /**
-     * 사용자가 칸을 이동할 때 사용하는 메서드
-     * <p>
-     * 이동을 위해 필요한 메서드의 반환 타입(return type), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-     */
     public MapStateDto move(BridgeCell userAnswer) {
         validate();
         int playerPosition = gameStates.getPlayerPosition();
@@ -37,14 +33,13 @@ public class BridgeGame {
         return MapStateDto.from(mapState);
     }
 
-    /**
-     * 사용자가 게임을 다시 시도할 때 사용하는 메서드
-     * <p>
-     * 재시작을 위해 필요한 메서드의 반환 타입(return type), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-     */
     public void retry() {
         this.gameStates.reset();
         this.mapState.reset();
+    }
+
+    public boolean isGameOver() {
+        return this.gameStates.isGameOver();
     }
